--- tests_and_analysis/test/euphonic_test/test_force_constants.py
@@ -1,1405 +1,0 @@
-from euphonic import ForceConstants, Crystal
-import pytest
-import os
-from ..utils import get_data_path
-import numpy.testing as npt
-import numpy as np
-import json
-from tests_and_analysis.test.euphonic_test.test_crystal import (
-    check_crystal_attrs
-)
-from slugify import slugify
-from collections import namedtuple
-from euphonic import ureg
-from contextlib import contextmanager
-import itertools
-
-ConstructorArgs = namedtuple(
-    "ConstructorArgs",
-    [
-        "crystal", "force_constants", "sc_matrix",
-        "cell_origins", "born", "dielectric"
-    ]
-)
-
-ExpectedData = namedtuple(
-    "ExpectedData", ["fc_mat_cell0_i0_j0", "fc_mat_celln"]
-)
-
-
-class ExpectedForceConstants:
-
-    def __init__(self, dirpath: str, with_material: bool = True):
-        """
-        Collect data from files for comparison to real ForceConstants files.
-
-        Parameters
-        ----------
-        dirpath : str
-            The directory path containing the files with the force constants
-            data in.
-        with_material : bool
-            True if the material is a prefix to the filename.
-        """
-        self.dirpath = dirpath
-        self.with_material = with_material
-        # Set up force constants
-        self._force_constants_unit = json.load(
-            open(self._get_file_from_dir_and_property("properties", "json"))
-        )["force_constants_unit"]
-        self._force_constants_mag = np.load(
-            self._get_file_from_dir_and_property("force_constants", "npy"),
-            allow_pickle=True
-        )
-        # Set up born
-        self._born_unit = json.load(
-            open(self._get_file_from_dir_and_property("properties", "json"))
-        )["born_unit"]
-        self._born_mag = np.load(
-            self._get_file_from_dir_and_property("born", "npy"),
-            allow_pickle=True
-        )
-        # Set up dielectric
-        self._dielectric_unit = json.load(
-            open(self._get_file_from_dir_and_property("properties", "json"))
-        )["dielectric_unit"]
-        self._dielectric_mag = np.load(
-            self._get_file_from_dir_and_property("dielectric", "npy"),
-            allow_pickle=True
-        )
-        # Set up sc_matrix
-        self.sc_matrix = np.load(
-            self._get_file_from_dir_and_property("sc_matrix", "npy"),
-            allow_pickle=True
-        )
-        # Set up cell origins
-        self.cell_origins = np.load(
-            self._get_file_from_dir_and_property("cell_origins", "npy"),
-            allow_pickle=True
-        )
-        # Set up crystal
-        self.crystal = Crystal.from_json_file(
-            self._get_file_from_dir_and_property("crystal", "json")
-        )
-        # Set up n_cells_in_sc
-        self.n_cells_in_sc = json.load(
-            open(self._get_file_from_dir_and_property("properties", "json"))
-        )["n_cells_in_sc"]
-
-    def _get_file_from_dir_and_property(self, property_name: str,
-                                        extension: str) -> str:
-        material = os.path.split(self.dirpath)[-1].lower()
-        if self.with_material:
-            filepath = os.path.join(
-                self.dirpath, material + "_" + property_name + "." + extension
-            )
-        else:
-            filepath = os.path.join(
-                self.dirpath, property_name + "." + extension
-            )
-        for f in os.listdir(self.dirpath):
-            f_path = os.path.join(self.dirpath, f)
-            if filepath == f_path:
-                return f_path
-        else:
-            raise FileNotFoundError("Could not find " + filepath)
-
-    @property
-    def born(self):
-        if self._born_mag.shape != ():
-            return self._born_mag * ureg(self._born_unit)
-        else:
-            return None
-
-    @property
-    def force_constants(self):
-        if self._force_constants_mag.shape != ():
-            return self._force_constants_mag * ureg(self._force_constants_unit)
-        else:
-            return None
-
-    @property
-    def dielectric(self):
-        if self._dielectric_mag.shape != ():
-            return self._dielectric_mag * ureg(self._dielectric_unit)
-        else:
-            return None
-
-    def to_dict(self):
-        d = {
-            'crystal': self.crystal.to_dict(),
-            'force_constants': self.force_constants.magnitude,
-            'force_constants_unit': str(self.force_constants.units),
-            'n_cells_in_sc': self.n_cells_in_sc,
-            'sc_matrix': self.sc_matrix,
-            'cell_origins': self.cell_origins,
-        }
-        if self.born is not None:
-            d["born"] = self.born.magnitude
-            d["born_unit"] = str(self.born.units)
-        if self.dielectric is not None:
-            d["dielectric"] = self.dielectric.magnitude
-            d["dielectric_unit"] = str(self.dielectric.units)
-        return d
-
-    def to_constructor_args(self, crystal=None, force_constants=None,
-                            sc_matrix=None, cell_origins=None,
-                            born=None, dielectric=None):
-        if crystal is None:
-            crystal = self.crystal
-        if force_constants is None:
-            force_constants = self.force_constants
-        if sc_matrix is None:
-            sc_matrix = self.sc_matrix
-        if cell_origins is None:
-            cell_origins = self.cell_origins
-        if born is None:
-            born = self.born
-        if dielectric is None:
-            dielectric = self.dielectric
-        return ConstructorArgs(
-            crystal, force_constants, sc_matrix, cell_origins, born, dielectric
-        )
-
-
-def check_force_constant_attrs(actual_force_constants, expected_force_constants,
-                               crystal=None, force_constants=None,
-                               sc_matrix=None, cell_origins=None,
-                               born=None, born_unit=None,
-                               dielectric=None, dielectric_unit=None,
-                               n_cells_in_sc=None, force_constants_unit=None):
-    check_fc_force_constants(
-        actual_force_constants, expected_force_constants,
-        force_constants, force_constants_unit
-    )
-    check_fc_sc_matrix(
-        actual_force_constants, expected_force_constants, sc_matrix
-    )
-    check_fc_cell_origins(
-        actual_force_constants, expected_force_constants, cell_origins
-    )
-    check_fc_born(
-        actual_force_constants, expected_force_constants,
-        born, born_unit
-    )
-    check_fc_dielectric(
-        actual_force_constants, expected_force_constants,
-        dielectric, dielectric_unit
-    )
-    check_fc_n_cells_in_sc(
-        actual_force_constants, expected_force_constants, n_cells_in_sc
-    )
-    check_fc_crystal(actual_force_constants, expected_force_constants, crystal)
-
-
-def check_fc_force_constants(actual_force_constants, expected_force_constants,
-                             overriding_force_constants=None,
-                             overriding_force_constants_unit=None):
-    # Allow override of expected force constants
-    force_constants = overriding_force_constants if \
-        overriding_force_constants is not None \
-        else expected_force_constants.force_constants
-    # Test values are sufficiently close
-    npt.assert_allclose(
-        actual_force_constants.force_constants.magnitude,
-        force_constants.magnitude
-    )
-    # Allow override of units
-    force_constants_unit = overriding_force_constants_unit if \
-        overriding_force_constants_unit is not None \
-        else expected_force_constants.force_constants.units
-    # Test unit matches
-    assert actual_force_constants.force_constants_unit == force_constants_unit
-
-
-def check_fc_sc_matrix(actual_force_constants, expected_force_constants,
-                       overriding_sc_matrix=None):
-    # Allow override
-    sc_matrix = overriding_sc_matrix if overriding_sc_matrix is not None \
-        else expected_force_constants.sc_matrix
-    # Test values are sufficiently close
-    npt.assert_allclose(actual_force_constants.sc_matrix, sc_matrix)
-
-
-def check_fc_cell_origins(actual_force_constants, expected_force_constants,
-                          overriding_cell_origins=None):
-    # Allow override
-    cell_origins = overriding_cell_origins if \
-        overriding_cell_origins is not None \
-        else expected_force_constants.cell_origins
-    # Test values are sufficiently close
-    npt.assert_allclose(actual_force_constants.cell_origins, cell_origins)
-
-
-def check_fc_crystal(actual_force_constants, expected_force_constants,
-                     overriding_crystal=None):
-    # Allow override
-    crystal = overriding_crystal if overriding_crystal is not None \
-        else expected_force_constants.crystal
-    # Defer testing of Crystal
-    check_crystal_attrs(
-        actual_force_constants.crystal, crystal
-    )
-
-
-def check_fc_n_cells_in_sc(actual_force_constants, expected_force_constants,
-                           overriding_n_cells_in_sc=None):
-    # Allow override
-    n_cells_in_sc = overriding_n_cells_in_sc if \
-        overriding_n_cells_in_sc is not None \
-        else expected_force_constants.n_cells_in_sc
-    # Test
-    assert actual_force_constants.n_cells_in_sc == n_cells_in_sc
-
-
-def check_fc_dielectric(actual_force_constants, expected_force_constants,
-                        overriding_dielectric=None,
-                        overriding_dielectric_unit=None):
-    # Allow override
-    dielectric = overriding_dielectric if overriding_dielectric is not None \
-        else expected_force_constants.dielectric
-    # dielectric is optional, detect if option has data in
-    if actual_force_constants.dielectric is not None:
-        npt.assert_allclose(
-            actual_force_constants.dielectric.magnitude, dielectric.magnitude
-        )
-        # Allow override
-        dielectric_unit = overriding_dielectric_unit if \
-            overriding_dielectric_unit is not None \
-            else expected_force_constants.dielectric.units
-        assert actual_force_constants.dielectric_unit == dielectric_unit
-    else:
-        # Assert that dielectric also is none in expected
-        assert dielectric is None
-
-
-def check_fc_born(actual_force_constants, expected_force_constants,
-                  overriding_born=None, overriding_born_unit=None):
-    # Allow override
-    born = overriding_born if overriding_born is not None \
-        else expected_force_constants.born
-    # born is optional, detect if option has data in
-    if actual_force_constants.born is not None:
-        npt.assert_allclose(
-            actual_force_constants.born.magnitude,
-            born.magnitude
-        )
-        # Allow override
-        born_unit = overriding_born_unit if overriding_born_unit is not None \
-            else born.units
-        assert actual_force_constants.born_unit == born_unit
-    else:
-        # Actual is none, check expected is the same
-        assert born is None
-
-
-def quartz_attrs():
-    test_data_dir = os.path.join(
-        get_data_path(), "interpolation", "quartz"
-    )
-    return ExpectedForceConstants(test_data_dir)
-
-
-@contextmanager
-def no_exception():
-    try:
-        yield
-    except Exception as e:
-        raise pytest.fail("Exception raised: " + str(e))
-
-
-@pytest.mark.unit
-class TestObjectCreation:
-
-    @pytest.mark.parametrize(("castep_bin_dir", "castep_bin_file"), [
-        ('LZO', 'La2Zr2O7.castep_bin'),
-        ('graphite', 'graphite.castep_bin'),
-        ('quartz', 'quartz.castep_bin')
-    ])
-    def test_creation_from_castep(self, castep_bin_dir, castep_bin_file):
-        dirpath = os.path.join(
-            get_data_path(), 'interpolation', castep_bin_dir
-        )
-        filepath = os.path.join(dirpath,  castep_bin_file)
-        fc = ForceConstants.from_castep(filepath)
-        check_force_constant_attrs(fc, ExpectedForceConstants(dirpath))
-
-    @pytest.mark.parametrize(("json_dir", "json_file"), [
-        ('LZO', 'lzo_force_constants.json'),
-        ('graphite', 'graphite_force_constants.json'),
-        ('quartz', 'quartz_force_constants.json')
-    ])
-    def test_creation_from_json(self, json_dir, json_file):
-        dirpath = os.path.join(
-            get_data_path(), 'interpolation', json_dir
-        )
-        filepath = os.path.join(dirpath, json_file)
-        fc = ForceConstants.from_json_file(filepath)
-        check_force_constant_attrs(fc, ExpectedForceConstants(dirpath))
-
-    @pytest.mark.parametrize(("json_dir", "json_file"), [
-        ('LZO', 'lzo_force_constants.json'),
-        ('graphite', 'graphite_force_constants.json'),
-        ('quartz', 'quartz_force_constants.json')
-    ])
-    def test_creation_from_dict(self, json_dir, json_file):
-        dirpath = os.path.join(
-            get_data_path(), 'interpolation', json_dir
-        )
-        expected_fc = ExpectedForceConstants(dirpath)
-        fc = ForceConstants.from_dict(expected_fc.to_dict())
-        check_force_constant_attrs(fc, expected_fc)
-
-    @pytest.mark.parametrize("phonopy_args", [
-        {"summary_name": "phonopy.yaml"},
-        {"summary_name": "phonopy_prim.yaml"},
-        {"summary_name": "phonopy_nofc.yaml", "fc_name": "FORCE_CONSTANTS"},
-        {"summary_name": "phonopy_nofc.yaml","fc_name": "FULL_FORCE_CONSTANTS"},
-        {
-            "summary_name": "phonopy_prim_nofc.yaml",
-            "fc_name": "PRIMITIVE_FORCE_CONSTANTS"
-        },
-        {
-            "summary_name": "phonopy_nofc.yaml",
-            "fc_name": "full_force_constants.hdf5"
-        },
-        {
-            "summary_name": "phonopy_nofc.yaml",
-            "fc_name": "force_constants.hdf5"
-        },
-        {"summary_name": "phonopy_nofc_noborn.yaml", "born_name": "BORN"},
-        {
-            "summary_name": "phonopy_prim_nofc.yaml",
-            "fc_name": "primitive_force_constants.hdf5"
-        }
-    ])
-    def test_creation_from_phonopy(self, phonopy_args):
-        test_data_dir = slugify(
-            "-".join(phonopy_args.keys()) + "-"
-            + "-".join(phonopy_args.values())
-        )
-        phonopy_args["path"] = os.path.join(
-            get_data_path(), 'phonopy_data', 'NaCl', 'interpolation'
-        )
-        fc = ForceConstants.from_phonopy(**phonopy_args)
-        expected_dirpath = os.path.join(phonopy_args["path"], test_data_dir)
-        check_force_constant_attrs(
-            fc, ExpectedForceConstants(expected_dirpath, with_material=False)
-        )
-
-    correct_elements = [
-        (
-            {
-                "born": quartz_attrs().to_constructor_args().born,
-                "dielectric": None
-            }
-        ),
-        (
-            {
-                "born": None,
-                "dielectric": quartz_attrs().to_constructor_args().dielectric
-            }
-        )
-    ]
-
-    @pytest.fixture(params=correct_elements)
-    def inject_elements(self, request):
-        injected_args = request.param
-        test_data_dir = os.path.join(
-            get_data_path(), "interpolation", "quartz"
-        )
-        expected_fc = ExpectedForceConstants(test_data_dir)
-        # Inject the faulty value and get a tuple of constructor arguments
-        args = expected_fc.to_constructor_args(**injected_args)
-        return args, expected_fc
-
-    def test_correct_object_creation(self, inject_elements):
-        args, expected_fc = inject_elements
-        fc = ForceConstants(*args)
-        check_force_constant_attrs(fc, expected_fc, **args._asdict())
-
-    faulty_elements = [
-        (
-            {
-                "sc_matrix": quartz_attrs().sc_matrix[:2]
-            },
-            ValueError
-        ),
-        (
-            {
-                "cell_origins": quartz_attrs().sc_matrix
-            },
-            ValueError
-        ),
-        (
-            {
-                "force_constants": quartz_attrs().force_constants[:2]
-            },
-            ValueError
-        ),
-        (
-            {
-                "born": quartz_attrs().born[:2]
-            },
-            ValueError
-        ),
-        (
-            {
-                "dielectric": quartz_attrs().dielectric[:2]
-            },
-            ValueError
-        ),
-        (
-            {
-                "sc_matrix": list(quartz_attrs().sc_matrix)
-            },
-            TypeError
-        ),
-        (
-            {
-                "cell_origins": quartz_attrs().sc_matrix * ureg.meter
-            },
-            TypeError
-        ),
-        (
-            {
-                "crystal": quartz_attrs().crystal.to_dict()
-            },
-            TypeError
-        ),
-        (
-            {
-                "force_constants": quartz_attrs().force_constants.magnitude
-            },
-            TypeError
-        ),
-        (
-            {
-                "born": list(quartz_attrs().born.magnitude)
-            },
-            TypeError
-        ),
-        (
-            {
-                "dielectric": quartz_attrs().dielectric.shape
-            },
-            TypeError
-        ),
-    ]
-
-    @pytest.fixture(params=faulty_elements)
-    def inject_faulty_elements(self, request):
-        faulty_args, expected_exception = request.param
-        test_data_dir = os.path.join(
-            get_data_path(), "interpolation", "quartz"
-        )
-        expected_fc = ExpectedForceConstants(test_data_dir)
-        # Inject the faulty value and get a tuple of constructor arguments
-        args = expected_fc.to_constructor_args(**faulty_args)
-        return args, expected_exception
-
-    def test_faulty_object_creation(self, inject_faulty_elements):
-        faulty_args, expected_exception = inject_faulty_elements
-        with pytest.raises(expected_exception):
-            ForceConstants(*faulty_args)
-
-    @pytest.fixture(params=[
-        (
-            'LZO', 'La2Zr2O7.castep_bin',
-            np.array([
-                [1.26492555e-01, -2.31204635e-31, -1.16997352e-13],
-                [-2.31204635e-31, 1.26492555e-01, 3.15544362e-30],
-                [-1.16997352e-13, 1.05181454e-30, 1.26492555e-01]
-            ]) * ureg('hartree/bohr**2'),
-            [3, 30, 33, 15, 18],
-            np.array([
-                [-8.32394989e-04, -2.03285211e-03, 3.55359333e-04],
-                [-2.22156212e-03, -6.29315975e-04, 1.21568713e-03],
-                [7.33617499e-05, 1.16282999e-03, 5.22410338e-05]
-            ]) * ureg('hartree/bohr**2')
-        ),
-        (
-            'graphite', 'graphite.castep_bin',
-            np.array([
-                [6.35111387e-01, 2.76471554e-18, 0.00000000e+00],
-                [2.05998413e-18, 6.35111387e-01, 0.00000000e+00],
-                [0.00000000e+00, 0.00000000e+00, 1.52513691e-01]
-            ]) * ureg('hartree/bohr**2'),
-            [10, 6, 9, 9, 12],
-            np.array([
-                [-8.16784177e-05, -1.31832252e-05, -1.11904290e-05],
-                [-1.31832252e-05, 5.42461675e-05, 3.73913780e-06],
-                [-1.11904290e-05, 3.73913780e-06, -2.99013850e-05]
-            ]) * ureg('hartree/bohr**2')
-        ),
-        (
-            'quartz', 'quartz.castep_bin',
-            np.array([
-                [0.22324308, 0.29855096, 0.31240272],
-                [0.29855096, 0.43968813, 0.34437270],
-                [0.31240272, 0.34437270, 0.33448280]
-            ]) * ureg('hartree/bohr**2'),
-            [4, 6, 9, 21, 24],
-            np.array([
-                [6.37874988e-05, -2.73116279e-04, -1.73624413e-04],
-                [1.99404810e-08, 4.40511248e-04, 2.33322837e-04],
-                [-1.85830030e-05, -9.05395392e-05, 2.17526152e-05]
-            ]) * ureg('hartree/bohr**2')
-        )
-    ])
-    def fc_mat(self, request):
-        directory, filename, fc_mat_cell0_i0_j0, celln, fc_mat_celln = \
-            request.param
-        path = os.path.join(get_data_path(), 'interpolation', directory)
-        expected_data = ExpectedData(
-            fc_mat_cell0_i0_j0=fc_mat_cell0_i0_j0,
-            fc_mat_celln=fc_mat_celln
-        )
-        filepath = os.path.join(path, filename)
-        fc = ForceConstants.from_castep(filepath)
-        return fc, celln, expected_data
-
-    def test_fc_mat_cells(self, fc_mat):
-        fc, celln, expected_data = fc_mat
-        npt.assert_allclose(
-            fc.force_constants[0, 0:3, 0:3].magnitude,
-            expected_data.fc_mat_cell0_i0_j0.magnitude
-        )
-        npt.assert_allclose(
-            fc.force_constants[
-                celln[0], celln[1]:celln[2], celln[3]:celln[4]
-            ].magnitude,
-            expected_data.fc_mat_celln.magnitude
-        )
-
-
-@pytest.mark.unit
-class TestInterpolatePhonons:
-
-    path = os.path.join(get_data_path(), 'interpolation')
-    unique_sc_offsets = [[] for _ in range(3)]
-    unique_cell_origins = [[] for _ in range(3)]
-
-    graphite_unique_sc_i = np.loadtxt(os.path.join(
-        path, 'graphite', 'graphite_unique_sc_i.txt'), dtype=np.int32)
-    graphite_unique_cell_i = np.loadtxt(os.path.join(
-        path, 'graphite', 'graphite_unique_cell_i.txt'), dtype=np.int32)
-    lzo_unique_sc_i = np.loadtxt(os.path.join(
-        path, 'LZO', 'lzo_unique_sc_i.txt'), dtype=np.int32)
-    lzo_unique_cell_i = np.loadtxt(os.path.join(
-        path, 'LZO', 'lzo_unique_cell_i.txt'), dtype=np.int32)
-
-    @pytest.fixture(params=[
-        (
-            'LZO', 'La2Zr2O7.castep_bin',
-            np.array([-1, 9.35, 3.35]),
-            np.loadtxt(
-                os.path.join(path, 'LZO', 'lzo_sc_phases.txt'),
-                dtype=np.complex128
-            ),
-            np.loadtxt(
-                os.path.join(path, 'LZO', 'lzo_cell_phases.txt'),
-                dtype=np.complex128
-            )
-        ),
-        (
-            'LZO', 'La2Zr2O7.castep_bin',
-            np.array([0.0, 0.0, 0.0]),  # At gamma point
-            np.full(
-                len(lzo_unique_sc_i), 1.0 + 0.0 * 1j, dtype=np.complex128
-            ),
-            np.full(
-                len(lzo_unique_cell_i), 1.0 + 0.0 * 1j, dtype=np.complex128
-            )
-        ),
-        (
-            'graphite', 'graphite.castep_bin',
-            np.array([0.001949, 0.001949, 0.0]),
-            np.loadtxt(os.path.join(
-                path, 'graphite', 'graphite_sc_phases.txt'
-            ), dtype=np.complex128),
-            np.loadtxt(os.path.join(
-                path, 'graphite', 'graphite_cell_phases.txt'
-            ), dtype=np.complex128)
-        ),
-        (
-            'graphite', 'graphite.castep_bin',
-            np.array([0.0, 0.0, 0.0]),  # At gamma point
-            np.full(
-                len(graphite_unique_sc_i), 1.0 + 0.0 * 1j, dtype=np.complex128
-            ),
-            np.full(
-                len(graphite_unique_cell_i), 1.0 + 0.0 * 1j, dtype=np.complex128
-            )
-        )
-    ])
-    def calculate_phases(self, request):
-        material_name, castep_bin_file, qpt, expected_sc_phases, \
-            expected_cell_phases = request.param
-        lower_material = material_name.lower()
-        filename = os.path.join(self.path, material_name, castep_bin_file)
-        self.unique_sc_i = np.loadtxt(
-            os.path.join(
-                self.path, material_name,
-                '{}_unique_sc_i.txt'.format(lower_material)
-            ), dtype=np.int32
-        )
-        self.unique_cell_i = np.loadtxt(
-            os.path.join(
-                self.path, material_name,
-                '{}_unique_cell_i.txt'.format(lower_material)
-            ), dtype=np.int32
-        )
-        unique_offsets_filepath = os.path.join(
-            self.path, material_name,
-            '{}_unique_sc_offsets.txt'.format(lower_material)
-        )
-        with open(unique_offsets_filepath) as f:
-            for i in range(3):
-                self.unique_sc_offsets[i] = [int(x)
-                                             for x in f.readline().split()]
-        unique_cell_origins_filepath = os.path.join(
-            self.path, material_name,
-            '{}_unique_cell_origins.txt'.format(lower_material)
-        )
-        with open(unique_cell_origins_filepath) as f:
-            for i in range(3):
-                self.unique_cell_origins[i] = [int(x)
-                                               for x in f.readline().split()]
-        return ForceConstants.from_castep(filename), qpt, \
-            expected_sc_phases, expected_cell_phases
-
-    def test_calculate_phases(self, calculate_phases):
-        fc, qpt, expected_sc_phases, expected_cell_phases = calculate_phases
-        sc_phases, cell_phases = fc._calculate_phases(
-            qpt, self.unique_sc_offsets, self.unique_sc_i,
-            self.unique_cell_origins, self.unique_cell_i
-        )
-        npt.assert_allclose(sc_phases, expected_sc_phases)
-        npt.assert_allclose(cell_phases, expected_cell_phases)
-
-    # Supercell image calculation limit - 2 supercells in each direction
-    lim = 2
-    n_atoms = 4
-    n_cells_in_sc = 98
-
-    def get_ijn_from_image_data(self, image_data):
-        i = image_data[:, 0].astype(int)
-        j = image_data[:, 1].astype(int)
-        n = image_data[:, 2].astype(int)
-        return i, j, n
-
-    @pytest.fixture
-    def lzo_calculate_supercell_images_n_sc_images(self):
-        image_data = np.loadtxt(
-            os.path.join(self.path, 'LZO', 'lzo_n_sc_images.txt')
-        )
-        i, j, n = self.get_ijn_from_image_data(image_data)
-        expected_n_images = np.zeros((22, 88))
-        expected_n_images[i, j] = n
-        # After refactoring where the shape of n_sc_images was changed from
-        # (n_atoms, n_cells_in_sc*n_atoms) to (n_cells_in_sc, n_atoms, n_atoms),
-        # expected_n_images must be reshaped to ensure tests still pass
-        expected_n_images = np.transpose(np.reshape(
-            expected_n_images, (22, 4, 22)), axes=[1, 0, 2]
-        )
-        filename = os.path.join(self.path, 'LZO', 'La2Zr2O7.castep_bin')
-        return ForceConstants.from_castep(filename), "_n_sc_images", \
-            expected_n_images
-
-    @pytest.fixture
-    def graphite_calculate_supercell_images_n_sc_images(self):
-        image_data = np.loadtxt(
-            os.path.join(self.path, 'graphite', 'graphite_n_sc_images.txt')
-        )
-        i, j, n = self.get_ijn_from_image_data(image_data)
-        expected_n_images = np.zeros(
-            (self.n_atoms, self.n_atoms * self.n_cells_in_sc))
-        expected_n_images[i, j] = n
-        # After refactoring where the shape of n_sc_images was changed from
-        # (n_atoms, n_cells_in_sc*n_atoms) to (n_cells_in_sc, n_atoms, n_atoms),
-        # expctc_n_images must be reshaped to ensure tests still pass
-        expected_n_images = np.transpose(
-            np.reshape(
-                expected_n_images,
-                (self.n_atoms, self.n_cells_in_sc, self.n_atoms)
-            ), axes=[1, 0, 2]
-        )
-        filename = os.path.join(self.path, 'graphite', 'graphite.castep_bin')
-        return ForceConstants.from_castep(filename), "_n_sc_images", \
-            expected_n_images
-
-    @pytest.fixture
-    def lzo_calculate_supercell_images_sc_image_i(self):
-        image_data = np.loadtxt(
-            os.path.join(self.path, 'LZO', 'lzo_sc_image_i.txt')
-        )
-        i, j, n = self.get_ijn_from_image_data(image_data)
-        sc_i = image_data[:, 3].astype(int)
-        max_n = np.max(n) + 1
-        # size = n_atoms X n_atoms*n_cells_in_sc X max supercell images
-        expected_sc_image_i = np.full((22, 88, max_n), -1)
-        expected_sc_image_i[i, j, n] = sc_i
-        # After refactoring where the shape of sc_image_i was changed from
-        # (n_atoms, n_cells_in_sc*n_atoms, (2*lim + 1)**3) to
-        # (n_cells_in_sc, n_atoms, n_atoms, (2*lim + 1)**3),
-        # expctd_image_i must be reshaped to ensure tests still pass
-        expected_sc_image_i = np.transpose(
-            np.reshape(expected_sc_image_i, (22, 4, 22, max_n)),
-            axes=[1, 0, 2, 3]
-        )
-        filename = os.path.join(self.path, 'LZO', 'La2Zr2O7.castep_bin')
-        return ForceConstants.from_castep(filename), "_sc_image_i", \
-            expected_sc_image_i
-
-    @pytest.fixture
-    def graphite_calculate_supercell_images_sc_image_i(self):
-        image_data = np.loadtxt(
-            os.path.join(self.path, 'graphite', 'graphite_sc_image_i.txt')
-        )
-        i, j, n = self.get_ijn_from_image_data(image_data)
-        sc_i = image_data[:, 3].astype(int)
-        max_n = np.max(n) + 1
-        # size = n_atoms X n_atoms*n_cells_in_sc X max supercell images
-        expected_sc_image_i = np.full(
-            (self.n_atoms, self.n_atoms * self.n_cells_in_sc, max_n), -1)
-        expected_sc_image_i[i, j, n] = sc_i
-        # After refactoring where the shape of sc_image_i was changed from
-        # (n_atoms, n_cells_in_sc*n_atoms, (2*lim + 1)**3) to
-        # (n_cells_in_sc, n_atoms, n_atoms, (2*lim + 1)**3),
-        # expctc_image_i must be reshaped to ensure tests still pass
-        expected_sc_image_i = np.transpose(
-            np.reshape(
-                expected_sc_image_i,
-                (self.n_atoms, self.n_cells_in_sc,self.n_atoms, max_n)
-            ), axes=[1, 0, 2, 3]
-        )
-        filename = os.path.join(self.path, 'graphite', 'graphite.castep_bin')
-        return ForceConstants.from_castep(filename), "_sc_image_i", \
-            expected_sc_image_i
-
-    @pytest.mark.parametrize('calculate_supercell_images', [
-        pytest.lazy_fixture('lzo_calculate_supercell_images_n_sc_images'),
-        pytest.lazy_fixture('lzo_calculate_supercell_images_sc_image_i'),
-        pytest.lazy_fixture('graphite_calculate_supercell_images_n_sc_images'),
-        pytest.lazy_fixture('graphite_calculate_supercell_images_sc_image_i')
-    ])
-    def test_calculate_supercell_images(self, calculate_supercell_images):
-        fc, testing_attribute, expected_attribute_data = \
-            calculate_supercell_images
-        fc._calculate_supercell_images(self.lim)
-        npt.assert_equal(
-            getattr(fc, testing_attribute), expected_attribute_data
-        )
-
-
-@pytest.mark.integration
-class TestCalculateQPointPhononModes:
-
-    path = os.path.join(get_data_path(), 'interpolation')
-
-    expected_freqs = {
-        "LZO": {
-            "asr": np.array([
-                [0.0002964449, 0.0002964449, 0.0003208033,
-                 0.0003501419, 0.0003501419, 0.0003903141,
-                 0.0004972138, 0.0004972138, 0.0005372809,
-                 0.0005438643, 0.0005861166, 0.0005861166,
-                 0.0007103804, 0.0007103804, 0.0007331639,
-                 0.0007786131, 0.0007874376, 0.0007929211,
-                 0.0008126016, 0.0008354861, 0.0008354861,
-                 0.0009078731, 0.0009078731, 0.0010160378,
-                 0.0010264374, 0.0010554444, 0.0011528145,
-                 0.0012094888, 0.0012304278, 0.0012410548,
-                 0.0012410548, 0.0012564500, 0.0013664070,
-                 0.0013664070, 0.0014355566, 0.0014355566,
-                 0.0014576129, 0.0015442745, 0.0015442745,
-                 0.0015449039, 0.0015505652, 0.0015937746,
-                 0.0017167608, 0.0017828465, 0.0017828465,
-                 0.0018048096, 0.0018598080, 0.0018598080,
-                 0.0018726170, 0.0019193824, 0.0020786777,
-                 0.0020786777, 0.0022934801, 0.0024275754,
-                 0.0024275754, 0.0024850292, 0.0025000804,
-                 0.0025179345, 0.0025179345, 0.0025401087,
-                 0.0025550191, 0.0025550191, 0.0028191070,
-                 0.0033473173, 0.0033680501, 0.0033680501],
-                [-1.2527213902e-09, -1.2524650945e-09, -1.2522509615e-09,
-                 2.5186476888e-04, 2.5186476888e-04, 2.5186476888e-04,
-                 4.2115533128e-04, 4.2115533128e-04, 4.5919137201e-04,
-                 4.5919137201e-04, 4.5919137201e-04, 6.0460911494e-04,
-                 6.0460911494e-04, 6.0460911494e-04, 6.1121916807e-04,
-                 6.1121916807e-04, 6.1121916807e-04, 6.8315329115e-04,
-                 6.8315329115e-04, 8.9085325717e-04, 8.9085325717e-04,
-                 8.9085325717e-04, 1.0237996415e-03, 1.0237996415e-03,
-                 1.0237996415e-03, 1.1666034640e-03, 1.1744920636e-03,
-                 1.1744920636e-03, 1.1744920636e-03, 1.2802064794e-03,
-                 1.2802064794e-03, 1.2802064794e-03, 1.3122749877e-03,
-                 1.4159439055e-03, 1.4159439055e-03, 1.4159439055e-03,
-                 1.4813308344e-03, 1.4813308344e-03, 1.5084818310e-03,
-                 1.5084818310e-03, 1.5084818310e-03, 1.5957863393e-03,
-                 1.7192994802e-03, 1.7192994802e-03, 1.8119521571e-03,
-                 1.8119521571e-03, 1.8119521571e-03, 1.8609709896e-03,
-                 1.8609709896e-03, 1.8609709896e-03, 2.1913629570e-03,
-                 2.2033452584e-03, 2.2033452584e-03, 2.2033452584e-03,
-                 2.4420900293e-03, 2.4420900293e-03, 2.4420900293e-03,
-                 2.4754830417e-03, 2.4754830417e-03, 2.4754830417e-03,
-                 2.5106852083e-03, 2.5106852083e-03, 2.5106852083e-03,
-                 3.3517193438e-03, 3.3517193438e-03, 3.3517193438e-03]
-            ]) * ureg('hartree'),
-            "no_asr": np.array([
-                [0.0002964623, 0.0002964623, 0.0003208033, 0.000350174,
-                 0.000350174, 0.0003903141, 0.0004972179, 0.0004972179,
-                 0.0005372886, 0.0005438642, 0.0005861163, 0.0005861163,
-                 0.0007103807, 0.0007103807, 0.0007331935, 0.0007786131,
-                 0.0007874315, 0.0007929211, 0.0008126019, 0.0008354958,
-                 0.0008354958, 0.000907874, 0.000907874, 0.0010160402,
-                 0.0010264376, 0.0010554468, 0.0011528125, 0.0012094888,
-                 0.0012304238, 0.0012410492, 0.0012410492, 0.00125645,
-                 0.0013664066, 0.0013664066, 0.0014355603, 0.0014355603,
-                 0.0014576129, 0.0015442837, 0.0015442837, 0.0015449061,
-                 0.0015505634, 0.0015937746, 0.0017167601, 0.0017828448,
-                 0.0017828448, 0.001804811, 0.0018598084, 0.0018598084,
-                 0.0018726182, 0.0019193824, 0.0020786769, 0.0020786769,
-                 0.002293487, 0.0024275755, 0.0024275755, 0.0024850292,
-                 0.0025000804, 0.002517934, 0.002517934, 0.0025401063,
-                 0.0025550198, 0.0025550198, 0.0028191102, 0.0033473155,
-                 0.0033680501, 0.0033680501],
-                [1.2522582708e-05, 1.2522582708e-05, 1.2522582708e-05,
-                 2.5186476888e-04, 2.5186476888e-04, 2.5186476888e-04,
-                 4.2115533128e-04, 4.2115533128e-04, 4.5920462007e-04,
-                 4.5920462007e-04, 4.5920462007e-04, 6.0462274991e-04,
-                 6.0462274991e-04, 6.0462274991e-04, 6.1121916807e-04,
-                 6.1121916807e-04, 6.1121916807e-04, 6.8315329115e-04,
-                 6.8315329115e-04, 8.9089400855e-04, 8.9089400855e-04,
-                 8.9089400855e-04, 1.0238000223e-03, 1.0238000223e-03,
-                 1.0238000223e-03, 1.1666034640e-03, 1.1744920636e-03,
-                 1.1744920636e-03, 1.1744920636e-03, 1.2802064794e-03,
-                 1.2802064794e-03, 1.2802064794e-03, 1.3122749877e-03,
-                 1.4159439055e-03, 1.4159439055e-03, 1.4159439055e-03,
-                 1.4813308344e-03, 1.4813308344e-03, 1.5085078032e-03,
-                 1.5085078032e-03, 1.5085078032e-03, 1.5957863393e-03,
-                 1.7192994802e-03, 1.7192994802e-03, 1.8119544413e-03,
-                 1.8119544413e-03, 1.8119544413e-03, 1.8609709896e-03,
-                 1.8609709896e-03, 1.8609709896e-03, 2.1913629570e-03,
-                 2.2033465408e-03, 2.2033465408e-03, 2.2033465408e-03,
-                 2.4420900293e-03, 2.4420900293e-03, 2.4420900293e-03,
-                 2.4754830417e-03, 2.4754830417e-03, 2.4754830417e-03,
-                 2.5106852083e-03, 2.5106852083e-03, 2.5106852083e-03,
-                 3.3517193438e-03, 3.3517193438e-03, 3.3517193438e-03]
-            ]) * ureg('hartree')
-        },
-        "graphite": {
-            "asr": np.array([
-                [-1.4954416353e-09, -1.4920078435e-09, -1.4801148666e-09,
-                 1.8361944616e-04, 1.8361944616e-04, 4.3310694904e-04,
-                 4.0094292783e-03, 4.0298656188e-03, 7.0857269505e-03,
-                 7.0857269505e-03, 7.1044827434e-03, 7.1044827434e-03],
-                [5.1653999633e-06, 3.2069627188e-05, 5.1670788680e-05,
-                 1.8632833975e-04, 1.9060943285e-04, 4.3314392040e-04,
-                 4.0093920383e-03, 4.0298240624e-03, 7.0856315683e-03,
-                 7.0859645028e-03, 7.1043776174e-03, 7.1046906727e-03],
-                [2.1681099868e-03, 2.2111725605e-03, 2.8193685159e-03,
-                 2.8197548439e-03, 2.9049709255e-03, 2.9141514127e-03,
-                 6.0277000188e-03, 6.0316368591e-03, 6.0977873315e-03,
-                 6.1104625371e-03, 6.3230408313e-03, 6.3262959952e-03],
-                [7.6871083387e-04, 8.8171407452e-04, 2.1005257455e-03,
-                 2.1043480758e-03, 3.5723508265e-03, 3.5925792804e-03,
-                 3.7679565502e-03, 3.7804526590e-03, 6.6594311463e-03,
-                 6.6651464071e-03, 7.2007012704e-03, 7.2132353835e-03],
-                [1.4065426444e-04, 1.4065426444e-04, 1.4065426444e-04,
-                 1.4065426444e-04, 3.2305904460e-04, 3.2305904460e-04,
-                 4.0222682768e-03, 4.0222682768e-03, 7.1591503492e-03,
-                 7.1591503492e-03, 7.1591503492e-03, 7.1591503492e-03]
-            ]) * ureg('hartree'),
-            "no_asr": np.array([
-                [-1.4947813885e-05, -1.4947813885e-05, 1.7871973126e-05,
-                 1.8361944616e-04, 1.8361944616e-04, 4.3310694904e-04,
-                 4.0094292783e-03, 4.0298656188e-03, 7.0857269505e-03,
-                 7.0857269505e-03, 7.1044827434e-03, 7.1044827434e-03],
-                [1.8594596548e-05, 2.8377223801e-05, 4.9463914519e-05,
-                 1.8632830014e-04, 1.9060940094e-04, 4.3314392073e-04,
-                 4.0093920384e-03, 4.0298240625e-03, 7.0856315682e-03,
-                 7.0859645027e-03, 7.1043776174e-03, 7.1046906727e-03],
-                [2.1681107372e-03, 2.2111728203e-03, 2.8193681109e-03,
-                 2.8197547098e-03, 2.9049711018e-03, 2.9141516000e-03,
-                 6.0276999661e-03, 6.0316367945e-03, 6.0977872708e-03,
-                 6.1104623404e-03, 6.3230407709e-03, 6.3262959354e-03],
-                [7.6869781440e-04, 8.8171407111e-04, 2.1005260482e-03,
-                 2.1043510112e-03, 3.5723508344e-03, 3.5925812023e-03,
-                 3.7679565852e-03, 3.7804526859e-03, 6.6594311404e-03,
-                 6.6651464171e-03, 7.2007012512e-03, 7.2132353897e-03],
-                [1.4065426444e-04, 1.4065426444e-04, 1.4065426444e-04,
-                 1.4065426444e-04, 3.2305904460e-04, 3.2305904460e-04,
-                 4.0222682768e-03, 4.0222682768e-03, 7.1591503492e-03,
-                 7.1591503492e-03, 7.1591503492e-03, 7.1591503492e-03]
-            ]) * ureg('hartree')
-        }
-    }
-
-    materials = [
-        (
-            "LZO", 'La2Zr2O7.castep_bin',
-            np.array([  # Qpoints
-                [-1.00, 9.35, 3.35],
-                [-1.00, 9.00, 3.00]
-            ])
-        ),
-        (
-            "graphite", "graphite.castep_bin",
-            np.array([  # Qpoints
-                [0.00, 0.00, 0.00],
-                [0.001949, 0.001949, 0.00],
-                [0.50, 0.00, 0.00],
-                [0.25, 0.00, 0.00],
-                [0.00, 0.00, 0.50]
-            ])
-        )
-    ]
-
-    kwargs = [
-        {"use_c": False, "fall_back_on_python": True, "n_threads": 1},
-        {"use_c": True, "fall_back_on_python": False, "n_threads": 1},
-        {"use_c": True, "fall_back_on_python": False, "n_threads": 2},
-        {
-            "use_c": False, "fall_back_on_python": True,
-            "n_threads": 1, "asr": "realspace"
-        },
-        {
-            "use_c": True, "fall_back_on_python": False,
-            "n_threads": 1, "asr": "realspace"
-        },
-        {
-            "use_c": True, "fall_back_on_python": False,
-            "n_threads": 2, "asr": "realspace"
-        }
-    ]
-
-    @pytest.fixture(
-        params=list(itertools.product(materials, kwargs))
-    )
-    def calculate_qpoint_phonon_modes(self, request):
-        material, kwargs = request.param
-        material_name, castep_bin_file, qpts = material
-        filename = os.path.join(self.path, material_name, castep_bin_file)
-        return ForceConstants.from_castep(filename), qpts, material_name, kwargs
-
-    def test_calculate_qpoint_phonon_modes(self, calculate_qpoint_phonon_modes):
-        fc, qpts, material_name, kwargs = calculate_qpoint_phonon_modes
-        print(material_name)
-        qpoint_phonon_modes = fc.calculate_qpoint_phonon_modes(
-            qpts, **kwargs
-        )
-        if "asr" in kwargs:
-            test_expected_freqs = self.expected_freqs[material_name]["asr"]
-        else:
-            test_expected_freqs = self.expected_freqs[material_name]["no_asr"]
-        npt.assert_allclose(
-            qpoint_phonon_modes.frequencies.to('hartree').magnitude,
-            test_expected_freqs.to('hartree').magnitude,
-            atol=1e-10
-        )
-
-    @pytest.mark.parametrize(
-        ("material_name", "castep_bin_file", "fc_mat_asr_file"),
-        [
-            ('graphite', 'graphite.castep_bin', 'graphite_fc_mat_asr.npy'),
-            ('LZO', 'La2Zr2O7.castep_bin', 'lzo_fc_mat_asr.npy')
-        ]
-    )
-    def test_enforce_realspace_acoustic_sum_rule(
-            self, material_name, castep_bin_file, fc_mat_asr_file):
-        castep_bin_filepath = os.path.join(
-            self.path, material_name, castep_bin_file
-        )
-        fc = ForceConstants.from_castep(castep_bin_filepath)
-        expected_fc_mat = np.load(
-            os.path.join(self.path, material_name, fc_mat_asr_file)
-        )
-        fc_mat = fc._enforce_realspace_asr()
-        npt.assert_allclose(fc_mat, expected_fc_mat, atol=1e-18)
-
-    quart_expected_freqs = {
-        "asr_splitting": np.array([
-            [0.00000000, 0.00000000, 0.00000134, 0.00058293, 0.00058293,
-             0.00101620, 0.00117558, 0.00117559, 0.00155134, 0.00168918,
-             0.00172394, 0.00172394, 0.00200565, 0.00200565, 0.00209131,
-             0.00245470, 0.00316467, 0.00316467, 0.00360268, 0.00363298,
-             0.00363298, 0.00487608, 0.00487608, 0.00495076, 0.00525430,
-             0.00525431, 0.00565964],
-            [0.00021598, 0.00056611, 0.00056686, 0.00075777, 0.00083532,
-             0.00083554, 0.00092328, 0.00119699, 0.00119903, 0.00175523,
-             0.00182882, 0.00191025, 0.00191045, 0.00205709, 0.00205738,
-             0.00259496, 0.00259505, 0.00355706, 0.00358489, 0.00362570,
-             0.00362572, 0.00484913, 0.00492809, 0.00492815, 0.00505041,
-             0.00553678, 0.00553679],
-            [0.00000000, 0.00000000, 0.00000134, 0.00058293, 0.00058293,
-             0.00101620, 0.00117558, 0.00117559, 0.00155134, 0.00168918,
-             0.00172394, 0.00172394, 0.00200565, 0.00200565, 0.00209131,
-             0.00245470, 0.00316467, 0.00316467, 0.00360268, 0.00363298,
-             0.00363298, 0.00487608, 0.00487608, 0.00495076, 0.00525430,
-             0.00525431, 0.00565964],
-            [0.00000000, 0.00000000, 0.00000093, 0.00058293, 0.00058305,
-             0.00101620, 0.00117558, 0.00117967, 0.00155134, 0.00161358,
-             0.00172394, 0.00176372, 0.00200565, 0.00209027, 0.00209131,
-             0.00238641, 0.00316467, 0.00317281, 0.00354729, 0.00363298,
-             0.00367378, 0.00487608, 0.00490124, 0.00495076, 0.00524661,
-             0.00525431, 0.00563557],
-            [0.00029928, 0.00044784, 0.00045479, 0.00083284, 0.00085117,
-             0.00107950, 0.00122947, 0.00131602, 0.00142866, 0.00154594,
-             0.00166280, 0.00179836, 0.00189458, 0.00191540, 0.00205403,
-             0.00246941, 0.00278199, 0.00315087, 0.00370009, 0.00376137,
-             0.00380880, 0.00488793, 0.00490540, 0.00490617, 0.00526645,
-             0.00531057, 0.00540940],
-            [0.00000000, 0.00000000, 0.00000093, 0.00058293, 0.00058305,
-             0.0010162, 0.00117558, 0.00117967, 0.00155134, 0.00161358,
-             0.00172394, 0.00176372, 0.00200565, 0.00209027, 0.00209131,
-             0.00238641, 0.00316467, 0.00317281, 0.00354729, 0.00363298,
-             0.00367378, 0.00487608, 0.00490124, 0.00495076, 0.00524661,
-             0.00525431, 0.00563557],
-            [0.00000000, 0.00000000, 0.00000000, 0.00058293, 0.00058293,
-             0.0010162, 0.00117558, 0.00117559, 0.00155134, 0.00157718,
-             0.00172394, 0.00172394, 0.00200565, 0.00200565, 0.00209131,
-             0.00222481, 0.00316467, 0.00316467, 0.00352002, 0.00363298,
-             0.00363298, 0.00487608, 0.00487608, 0.00492353, 0.00495076,
-             0.0052543, 0.00525431],
-            [0.00000000, 0.00000000, 0.00000000, 0.00058293, 0.00058293,
-             0.0010162, 0.00117558, 0.00117559, 0.00155134, 0.00157718,
-             0.00172394, 0.00172394, 0.00200565, 0.00200565, 0.00209131,
-             0.00222481, 0.00316467, 0.00316467, 0.00352002, 0.00363298,
-             0.00363298, 0.00487608, 0.00487608, 0.00492353, 0.00495076,
-             0.0052543, 0.00525431],
-            [0.00000000, 0.00000000, 0.00000000, 0.00058293, 0.00058293,
-             0.0010162, 0.00117558, 0.00117559, 0.00155134, 0.00157718,
-             0.00172394, 0.00172394, 0.00200565, 0.00200565, 0.00209131,
-             0.00222481, 0.00316467, 0.00316467, 0.00352002, 0.00363298,
-             0.00363298, 0.00487608, 0.00487608, 0.00492353, 0.00495076,
-             0.0052543, 0.00525431],
-            [0.00000000, 0.00000000, 0.00000000, 0.00058293, 0.00058293,
-             0.0010162, 0.00117558, 0.00117559, 0.00155134, 0.00157718,
-             0.00172394, 0.00172394, 0.00200565, 0.00200565, 0.00209131,
-             0.00222481, 0.00316467, 0.00316467, 0.00352002, 0.00363298,
-             0.00363298, 0.00487608, 0.00487608, 0.00492353, 0.00495076,
-             0.0052543, 0.00525431],
-            [0.00000000, 0.00000000, 0.00000102, 0.00058293, 0.00058303,
-             0.0010162, 0.00117558, 0.00117884, 0.00155134, 0.00162384,
-             0.00172394, 0.00176016, 0.00200565, 0.00207171, 0.00209131,
-             0.00240211, 0.00316467, 0.0031712, 0.0035549, 0.00363298,
-             0.00366903, 0.00487608, 0.00489617, 0.00495076, 0.00524826,
-             0.00525431, 0.00564028],
-            [0.00028671, 0.00036112, 0.00049916, 0.00082386, 0.00084410,
-             0.00109225, 0.00130495, 0.00134356, 0.00142776, 0.00155337,
-             0.00159266, 0.00179031, 0.00188151, 0.00193612, 0.00205664,
-             0.00243853, 0.00280084, 0.00315578, 0.00370595, 0.00370979,
-             0.00385195, 0.00489429, 0.00489894, 0.00490641, 0.00527571,
-             0.00533034, 0.00537398],
-            [0.00000000, 0.00000000, 0.00000102, 0.00058293, 0.00058303,
-             0.0010162, 0.00117558, 0.00117884, 0.00155134, 0.00162384,
-             0.00172394, 0.00176016, 0.00200565, 0.00207171, 0.00209131,
-             0.00240211, 0.00316467, 0.0031712, 0.0035549, 0.00363298,
-             0.00366903, 0.00487608, 0.00489617, 0.00495076, 0.00524826,
-             0.00525431, 0.00564028]
-        ]) * ureg('hartree'),
-        "asr": np.array([
-            [0.00000000, 0.00000000, 0.00000000, 0.00058293, 0.00058293,
-             0.00101620, 0.00117558, 0.00117559, 0.00155134, 0.00157718,
-             0.00172394, 0.00172394, 0.00200565, 0.00200565, 0.00209131,
-             0.00222481, 0.00316467, 0.00316467, 0.00352002, 0.00363298,
-             0.00363298, 0.00487608, 0.00487608, 0.00492353, 0.00495076,
-             0.00525430, 0.00525431],
-            [0.00021598, 0.00056611, 0.00056686, 0.00075777, 0.00083532,
-             0.00083554, 0.00092328, 0.00119699, 0.00119903, 0.00175523,
-             0.00182882, 0.00191025, 0.00191045, 0.00205709, 0.00205738,
-             0.00259496, 0.00259505, 0.00355706, 0.00358489, 0.00362570,
-             0.00362572, 0.00484913, 0.00492809, 0.00492815, 0.00505041,
-             0.00553678, 0.00553679],
-            [0.00029928, 0.00044784, 0.00045479, 0.00083284, 0.00085117,
-             0.00107950, 0.00122947, 0.00131602, 0.00142866, 0.00154594,
-             0.00166280, 0.00179836, 0.00189458, 0.00191540, 0.00205403,
-             0.00246941, 0.00278199, 0.00315087, 0.00370009, 0.00376137,
-             0.00380880, 0.00488793, 0.00490540, 0.00490617, 0.00526645,
-             0.00531057, 0.00540940],
-            [0.00028671, 0.00036112, 0.00049916, 0.00082386, 0.00084410,
-             0.00109225, 0.00130495, 0.00134356, 0.00142776, 0.00155337,
-             0.00159266, 0.00179031, 0.00188151, 0.00193612, 0.00205664,
-             0.00243853, 0.00280084, 0.00315578, 0.00370595, 0.00370979,
-             0.00385195, 0.00489429, 0.00489894, 0.00490641, 0.00527571,
-             0.00533034, 0.00537398]
-        ]) * ureg('hartree'),
-        "no_asr": np.array([
-            [-0.00009745, -0.00005474, -0.00005474, 0.00058293, 0.00058293,
-             0.00101620, 0.00117558, 0.00117559, 0.00155134, 0.00157713,
-             0.00172393, 0.00172393, 0.00200563, 0.00200563, 0.00209131,
-             0.00222477, 0.00316467, 0.00316467, 0.00352001, 0.00363297,
-             0.00363297, 0.00487607, 0.00487607, 0.00492351, 0.00495076,
-             0.00525430, 0.00525431],
-            [0.00021293, 0.00056589, 0.00056590, 0.00075771, 0.00083485,
-             0.00083485, 0.00092258, 0.00119698, 0.00119698, 0.00175516,
-             0.00182881, 0.00191025, 0.00191027, 0.00205705, 0.00205705,
-             0.00259491, 0.00259491, 0.00355700, 0.00358489, 0.00362570,
-             0.00362570, 0.00484908, 0.00492808, 0.00492808, 0.00505040,
-             0.00553676, 0.00553677],
-            [0.00029664, 0.00044691, 0.00045286, 0.00083218, 0.00085090,
-             0.00107911, 0.00122936, 0.00131583, 0.00142863, 0.00154578,
-             0.00166267, 0.00179816, 0.00189441, 0.00191534, 0.00205378,
-             0.00246936, 0.00278196, 0.00315079, 0.00370003, 0.00376130,
-             0.00380875, 0.00488787, 0.00490533, 0.00490608, 0.00526644,
-             0.00531051, 0.00540936],
-            [0.00028396, 0.00036024, 0.00049599, 0.00082340, 0.00084375,
-             0.00109203, 0.00130479, 0.00134347, 0.00142754, 0.00155335,
-             0.00159243, 0.00179022, 0.00188140, 0.00193608, 0.00205639,
-             0.00243849, 0.00280079, 0.00315569, 0.00370591, 0.00370973,
-             0.00385188, 0.00489424, 0.00489891, 0.00490631, 0.00527567,
-             0.00533033, 0.00537392]
-        ]) * ureg('hartree')
-    }
-
-    qpts = np.array([
-        [0.00, 0.00, 0.00],
-        [0.00, 0.00, 0.50],
-        [-0.25, 0.50, 0.50],
-        [-0.151515, 0.575758, 0.5]
-    ])
-    split_qpts =  np.array([
-        [0.00, 0.00, 0.00],
-        [0.00, 0.00, 0.50],
-        [0.00, 0.00, 0.00],
-        [0.00, 0.00, 0.00],
-        [-0.25, 0.50, 0.50],
-        [0.00, 0.00, 0.00],
-        [0.00, 0.00, 0.00],
-        [0.00, 0.00, 0.00],
-        [0.00, 0.00, 0.00],
-        [0.00, 0.00, 0.00],
-        [0.00, 0.00, 0.00],
-        [-0.151515, 0.575758, 0.5],
-        [0.00, 0.00, 0.00]
-    ])
-    split_qpts_insert_gamma = np.array([
-        [0.00, 0.00, 0.00],
-        [0.00, 0.00, 0.50],
-        [0.00, 0.00, 0.00],
-        [-0.25, 0.50, 0.50],
-        [0.00, 0.00, 0.00],
-        [0.00, 0.00, 0.00],
-        [0.00, 0.00, 0.00],
-        [-0.151515, 0.575758, 0.5],
-        [0.00, 0.00, 0.00]
-    ])
-
-    quartz_test_data = [
-        (
-            {"qpts": qpts, "dipole": True, "splitting": False},
-            2e-6
-        ),
-        (
-            {
-                "qpts": qpts, "dipole": True, "splitting": False,
-                "use_c": True, "fall_back_on_python": False
-            },
-            2e-6
-        ),
-        (
-            {
-                "qpts": qpts, "dipole": True, "splitting": False, "use_c": True,
-                "n_threads": 2, "fall_back_on_python": False
-            },
-            8e-8
-        ),
-        (
-            {
-                "qpts": qpts, "asr": 'reciprocal',
-                "dipole": True, "splitting": False
-            },
-            5e-4
-        ),
-        (
-            {
-                "qpts": qpts, "asr": 'reciprocal', "dipole": True,
-                "splitting": False, "use_c": True, "fall_back_on_python": False
-            },
-            5e-4
-        ),
-        (
-            {
-                "qpts": qpts, "asr": 'reciprocal', "dipole": True,
-                "splitting": False,"use_c": True, "n_threads": 2,
-                "fall_back_on_python": False
-            },
-            5e-4
-        ),
-        (
-            {
-                "qpts": split_qpts, "asr": 'reciprocal',
-                "dipole": True, "splitting": True
-            },
-            5e-4
-        ),
-        (
-            {
-                "qpts": split_qpts, "asr": 'reciprocal', "dipole": True,
-                "splitting": True, "use_c": True, "fall_back_on_python": False
-            },
-            5e-4
-        ),
-        (
-            {
-                "qpts": split_qpts, "asr": 'reciprocal', "dipole": True,
-                "splitting": True, "use_c": True, "n_threads": 2,
-                "fall_back_on_python": False
-            },
-            5e-4
-        ),
-        (
-            {
-                "qpts": split_qpts_insert_gamma, "asr": 'reciprocal',
-                "dipole": True, "splitting": True, "insert_gamma": True
-            },
-            5e-4
-        )
-    ]
-
-    quartz_castep_bin_file = os.path.join(path, "quartz", "quartz.castep_bin")
-
-    @pytest.mark.parametrize(("kwargs", "atol"), quartz_test_data)
-    def test_quartz_calculate_qpoint_phonon_modes(self, kwargs, atol):
-        fc = ForceConstants.from_castep(self.quartz_castep_bin_file)
-        qpoint_phonon_modes = fc.calculate_qpoint_phonon_modes(**kwargs)
-        if "asr" in kwargs:
-            if "splitting" in kwargs and kwargs["splitting"]:
-                test_expected_freqs = self.quart_expected_freqs["asr_splitting"]
-            else:
-                test_expected_freqs = self.quart_expected_freqs["asr"]
-        else:
-            test_expected_freqs = self.quart_expected_freqs["no_asr"]
-        npt.assert_allclose(
-            qpoint_phonon_modes.frequencies.to('hartree').magnitude,
-            test_expected_freqs.to('hartree').magnitude,
-            atol=atol
-        )
-
-    phonopy_yaml_file = os.path.join(
-        get_data_path(), "phonopy_data", "NaCl", "interpolation"
-    )
-
-    phonopy_expected_freqs = {
-        "no_asr": np.array([
-            [1.68554227, 1.68554227, 1.68554227, 13.03834226, 13.03834226,
-             13.03834226, 13.70911614, 13.70911614, 13.70911614, 13.70911614,
-             13.70911614, 13.70911614, 14.77619334, 14.77619334, 14.77619334,
-             14.77619334, 14.77619334, 14.77619334, 15.29016584, 15.29016584,
-             15.29016584, 26.8034935, 26.8034935, 26.8034935],
-            [10.43378439, 10.43378439, 10.43378439, 10.43378439, 11.22623974,
-             11.22623974, 13.41709858, 13.41709858, 13.41709858, 13.41709858,
-             14.52964205, 14.52964205, 14.52964205, 14.52964205, 15.35508239,
-             15.35508239, 16.77483373, 16.77483373, 21.52053536, 21.52053536,
-             21.52053536, 21.52053536, 31.90031516, 31.90031516],
-            [9.70009213, 9.70009213, 9.70009213, 9.70009213, 12.23105123,
-             12.23105123, 12.23105123, 12.23105123, 12.89476859, 12.89476859,
-             12.89476859, 12.89476859, 13.13693891, 13.13693891, 13.13693891,
-             13.13693891, 22.06651219, 22.06651219, 22.06651219, 22.06651219,
-             29.1978153, 29.1978153, 29.1978153, 29.1978153],
-            [8.70234882, 8.70234882, 10.00872202, 10.00872202, 12.85563966,
-             12.85563966, 12.91640758, 12.91640758, 13.51524747, 13.51524747,
-             13.69887901, 13.69887901, 13.76367323, 13.76367323, 14.54189731,
-             14.54189731, 19.40187956, 19.40187956, 21.05182267, 21.05182267,
-             28.28975768, 28.28975768, 29.77808371, 29.77808371]
-        ]) * ureg('meV'),
-        "realsp_asr": np.array([
-            [-1.60758061, -1.60718395, -1.60708352, 13.25894361, 13.25895397,
-             13.25895862, 13.7067571, 13.7067572, 13.70675726, 13.70675728,
-             13.70675731, 13.70675743, 14.77838148, 14.77838151, 14.77838152,
-             14.77838152, 14.77838152, 14.77838155, 15.28987378, 15.2898738,
-             15.2898738, 26.80366009, 26.8036601, 26.8036601],
-            [10.42911169, 10.42911169, 10.42918151, 10.42918158, 11.25702209,
-             11.25705428, 13.42071458, 13.42071459, 13.42073099, 13.42073099,
-             14.5376376, 14.5376376, 14.5376376, 14.5376376, 15.35508239,
-             15.35508239, 16.77483373, 16.77483373, 21.51513497, 21.51513497,
-             21.51513497, 21.51513497, 31.88946119, 31.88946565],
-            [9.69770891, 9.69770893, 9.69770898, 9.69770898, 12.23182976,
-             12.23182984, 12.23182984, 12.23182986, 12.89656098, 12.89656098,
-             12.89656098, 12.89656098, 13.13865655, 13.13865656, 13.13865656,
-             13.13865658, 22.06696896, 22.066969, 22.066969, 22.06696901,
-             29.19637097, 29.19637097, 29.19637098, 29.19637098],
-            [8.70045642, 8.70045874, 10.00553973, 10.00554104, 12.85834764,
-             12.85834786, 12.91733905, 12.91734389, 13.51723245, 13.51723438,
-             13.70221457, 13.70221553, 13.76553799, 13.76553943, 14.54322351,
-             14.54322506, 19.40371223, 19.40371496, 21.0515387, 21.05154028,
-             28.28756337, 28.28756345, 29.77527494, 29.77527501]
-        ]) * ureg('meV'),
-        "recip_asr": np.array([
-            [-0.00001183, -0.00000838, 0.00000036, 13.0330645, 13.0330645,
-             13.0330645, 13.70911614, 13.70911614, 13.70911614, 13.70911614,
-             13.70911614, 13.70911614, 14.77619334, 14.77619334, 14.77619334,
-             14.77619334, 14.77619334, 14.77619334, 15.29016584, 15.29016584,
-             15.29016584, 26.8034935, 26.8034935, 26.8034935],
-            [10.29421474, 10.29421474, 10.43378439, 10.43378439, 11.09817998,
-             11.22623974, 13.41390631, 13.41390631, 13.41709858, 13.41709858,
-             14.52964205, 14.52964205, 14.52964205, 14.52964205, 15.35508239,
-             15.35508239, 16.77483373, 16.77483373, 21.52053536, 21.52053536,
-             21.52053536, 21.52053536, 31.89843751, 31.90031516],
-            [9.63825967, 9.63825967, 9.70009213, 9.70009213, 12.17415977,
-             12.20855918, 12.20855918, 12.23105123, 12.88310822, 12.88310822,
-             12.89476859, 12.89476859, 13.12198284, 13.13553123, 13.13553123,
-             13.13693891, 22.04350987, 22.05424945, 22.05424945, 22.06651219,
-             29.19174209, 29.19174209, 29.19468408, 29.1978153],
-            [8.65321026, 8.65334846, 9.98387815, 9.98421216, 12.84875298,
-             12.85137816, 12.86815343, 12.90260046, 13.51181836, 13.51450714,
-             13.69142224, 13.69831554, 13.76153287, 13.76224884, 14.51838141,
-             14.5334654, 19.37183479, 19.39110497, 21.03444816, 21.04652258,
-             28.28732511, 28.28804436, 29.77604856, 29.77629323]
-        ]) * ureg('meV')
-    }
-
-    @pytest.mark.parametrize(
-        ("kwargs"), [
-            {"dipole": True, "splitting": False},
-            {
-                "dipole": True, "splitting": False, "asr": None,
-                "use_c": True, "fall_back_on_python": False
-            },
-            {
-                "dipole": True, "splitting": False,
-                "use_c": True, "fall_back_on_python": False, "n_threads": 2
-            },
-            {"dipole": True, "splitting": False, "asr": 'reciprocal'},
-            {
-                "dipole": True, "splitting": False, "asr": 'reciprocal',
-                "use_c": True, "fall_back_on_python": False
-            },
-            {
-                "dipole": True, "splitting": False, "asr": 'reciprocal',
-                "use_c": True, "fall_back_on_python": False, "n_threads": 2
-            },
-            {"dipole": True, "splitting": False, "asr": 'realspace'},
-            {
-                "dipole": True, "splitting": False, "asr": 'realspace',
-                "use_c": True, "fall_back_on_python": False
-            },
-            {
-                "dipole": True, "splitting": False, "asr": 'realspace',
-                "use_c": True, "fall_back_on_python": False, "n_threads": 2
-            }
-        ]
-    )
-    def test_phonopy_calculate_qpoint_phonon_modes(self, kwargs):
-        fc = ForceConstants.from_phonopy(
-            path=self.phonopy_yaml_file, summary_name="phonopy.yaml"
-        )
-        if "asr" in kwargs and kwargs["asr"] == 'realspace':
-            expected_freqs = self.phonopy_expected_freqs["realsp_asr"]
-        elif "asr" in kwargs and kwargs["asr"] == "reciprocal":
-            expected_freqs = self.phonopy_expected_freqs["recip_asr"]
-        else:
-            expected_freqs = self.phonopy_expected_freqs["no_asr"]
-        qpts = np.array([
-            [0.00, 0.00, 0.00],
-            [0.00, 0.00, 0.50],
-            [-0.25, 0.50, 0.50],
-            [-0.151515, 0.575758, 0.5]
-        ])
-        qpoint_phonon_modes = fc.calculate_qpoint_phonon_modes(qpts, **kwargs)
-        npt.assert_allclose(
-            qpoint_phonon_modes.frequencies.to('hartree').magnitude,
-            expected_freqs.to('hartree').magnitude,
-            atol=1e-8
-        )
